eliminazione

+++ include/gamelogic/GameBoard.h
@@ -26,7 +26,7 @@ typedef enum {
 typedef class GameBoardElement {
     GameBoardElementType type;
     GameBoardElementColor color;
-    unsigned long id;   /* same id means merged pill, ignored for virus */
+    int id;   /* same id means merged pill, ignored for virus */
 } GameBoardElement;
 
 #endif
